change account unique constraint to battletag
Account battlenet_id assumed to be unique in the upstream API. It is not the case, the API returns the id of a character
rather than an account. There is no way to get the account id for the API, so removing it completely, making account cross
regional entity. Region info is moved to the player_character relation.