add full anonymization task
This task runs anonymization task for all accounts, and not only for the recent ones. This fixes a bug where anonymous accounts can be out of sync with their characters when some of their characters were  rebound to a new active account, and account's anonymous btag was based on that character, which could lead to duplicate unique constraint exception being thrown when new active account is anonymized again.