add redundant d-none style
This gives d-none class the highest priority, which overrides Bootstrap's d-* classes. This simplifies visibility handling when visibility is toggled by adding/removing of the d-none class.
This also fixes a bug where visibility is not properly handled because multiple d-* classes are present, which may override the visibility in some cases.